+++ src/widgets/push_button_widget.rs
@@ -26,6 +26,7 @@ use sdl2::video::Window;
 
 use crate::widgets::text_widget::{TextJustify, TextWidget};
 use sdl2::pixels::Color;
+use std::any::Any;
 use std::collections::HashMap;
 
 /// This is the callback type that is used when an `on_click` callback is triggered from this
@@ -186,6 +187,7 @@ impl Widget for PushButtonWidget {
         self.button_clicked_callback(_widgets, _button, _clicks, _state);
     }
 
+    default_widget_functions!();
     default_widget_properties!();
     default_widget_callbacks!();
 }